build_data_file {voteswitchR}	R Documentation

Harmonize, impute, map, rake, and aggregate vote switching data

Description

Runs a ShinyApp that guides users through the steps required for data download, recoding, mapping, and imputation.

Usage

build_data_file()

Value

Returns and/or stores an object named data_file which contains a list of data for all requested electoral contexts.

[Package voteswitchR version 0.1.0 Index]