calculate_pred_error {voteswitchR}	R Documentation

Calculate the MAEs and RMSEs of MAVCL Predictions

Description

Uses an object of class mavcl_est and calculates prediction errors of posterior means or medians.

Usage

calculate_pred_error(
  mavcl_object,
  posterior = c("mean", "median"),
  type = c("mae", "rmse"),
  re_null = FALSE
)

Arguments

mavcl_object
An object of class mavcl_est.

posterior
"mean" or "median", depending on whether posterior means or medians should be evaluated against the observed proportions of the outcome variable.

type
"mae" or "rmse", depending on whether (mean) average errors or the root mean squared standard error should be reported

re_null
logical, specifies whether the prediction should be calculated with (FALSE) or without (TRUE) random effects

Value

Returns a list with observation-/outcome-specific absolute errors and the overall mean absolute error (if type == "mae") or the overall root mean squared error (if type == "rmse")

[Package voteswitchR version 0.1.0 Index]