compute_qoi {voteswitchR}	R Documentation

Compute MAVCL quantities of interest

Description

Computes quantities of interest such as conditional expectations and average marginal effects/first differences from a MAVCL estimation object.

Usage

compute_qoi(
  mavcl_object,
  y_structure,
  posterior_quantiles = c(0.5, 0.025, 0.975),
  full_posterior = FALSE,
  len_continuous_sequence = 21L,
  conditional_expectation = TRUE,
  average_marginal_effect = TRUE,
  ame_shift = "tiny",
  base = c("t", "tm1", "avg"),
  relative = TRUE,
  atmeans = FALSE,
  re_null = FALSE
)

Arguments

mavcl_object
An estimation object of class mavcl_est.

y_structure
Structure of switching patterns, as generated by voteswitchR::recode_switches().

posterior_quantiles
A numeric vector of quantile values for the posterior quantile summaries.

full_posterior
Logical; if TRUE, the full posterior draws will be return alongside the posterior quantile summaries.

len_continuous_sequence
Suggested length of the value sequence for continuous predictors and/or moderators.

conditional_expectation
Logical; if TRUE, conditional expectations (a.k.a. "expected values" or "predicitions") will be returned.

average_marginal_effect
Logical; if TRUE, average marginal effects (continuous main predictor) or average first difference (categorical main predictor) will be returned.

base
Specifies whether the denominator for the normalization of relative quantities of interest. Can be "t" for parties' electorates at t, "tm1" for parties' electorates at t - 1, or "avg" for parties' average inter-election electorates.

relative
Logical; if TRUE, the returned quantities of interest are normalized by base.

atmeans
Logical; if TRUE, quantities of interest will be calculated for an average observation. If FALSE, quantities will be calculated at the observed values of all observations and then averaged.

re_null
Logical; if TRUE,only the "fixed" portion of the model coefficients will be used and the "random" portion discarded.

Value

Returns a nested list of class voteswitchR_qoi that contains the requested quantities.

[Package voteswitchR version 0.1.0 Index]