data_guide {voteswitchR}	R Documentation

Versions, download links, and access details for survey data

Description

A detailed documentation of the survey data that voteswitchR's build_data_file function processes.

Usage

data_guide

Format

A data frame with 220 rows and 7 columns.

[Package voteswitchR version 0.1.0 Index]