recode_switches {voteswitchR}	R Documentation

Recode aggregate switching patterns

Description

Recodes raw aggregate switching patterns according to a user-supplied factor variable included in a (possibly user-supplied) version of mappings.

Usage

recode_switches(
  switches,
  subgroup = NULL,
  stack_by_subgroup = FALSE,
  assign_others = NULL,
  mappings = NULL,
  switch_factor,
  type = c("elections", "party-elections")
)

Arguments

switches
Aggregate counts of switches as generated by voteswitchR::build_data_file().

subgroup
A character that specifies the column-name of electorate subgroups. Should be NULL (the default) if no subgroup analyses are required.

stack_by_subgroup
Logical; if TRUE, subgroups are stacked into long format such that switching proportions from the MAVCL model sum to one within subgroups. If FALSE, subgroup-specific switching is recorded in wide format such that switching proportions for each subgroup sum to the size of the subgroup. Only takes effect if subgroup != NULL.

assign_others
Optional: A character that specifies a level for residual other parties for which switch_factor information is not available.

mappings
A copy of voteswitchR::mappings, possibly enriched with external party and/or election-specific variables.

switch_factor
Column name of a factor variable in mappings that specifies the marginal categories of the desired voter transition matrix.

type
Specifies whether to return election-specific voter-transition matrices ("elections") or stacked party-within-election transition matrices ("party-elections"). The latter should be used for studying the effects of party strategies.

Value

Returns the supplied data frame mappings, extended by columns for each switching pattern detected in the data.

[Package voteswitchR version 0.1.0 Index]