run_mavcl {voteswitchR}	R Documentation

Run the MAVCL model

Description

Runs the Mixed Aggregate Varying Choice Set Logit (MAVCL) model via rstan. By default, the model includes election-cell-specific intercepts (i.e., random intercepts for each cell of an election-specific voter transition matrix).

Usage

run_mavcl(
  data,
  y_names,
  null_model = FALSE,
  main_predictor,
  predictor_continuous = TRUE,
  moderator = NULL,
  moderator_continuous = TRUE,
  other_covariates = NULL,
  random_slopes = FALSE,
  re_parties = FALSE,
  re_elections = FALSE,
  re_countries = FALSE,
  iter = 2000L,
  warmup = 1000L,
  thin = 2L,
  chains = 2L,
  cores = parallel::detectCores(),
  max_treedepth = 15,
  adapt_delta = 0.8,
  seed,
  parallelize = FALSE,
  savename = "mavcl"
)

Arguments

data
A data object generated via voteswitchR::build_data_file() and processed via voteswitchR::recode_switches(). Can be a single tbl_df/tbl/data.frame or, in the case multiply imputed vote switching counts, a list thereof.

y_names
A vector of column names for vote switching counts, as generated by voteswitchR::recode_switches.

null_model
Logical; if TRUE, the function runs a null model without any right-hand side variables.

main_predictor
Column name of the main predictor of interest.

predictor_continuous
Logical; if TRUE, the predictor is treated as continuous; if FALSE, it is treated as a factor.

moderator
Column name of a moderator to be interacted with the main predictor or NULL if there is no moderator.

moderator_continuous
Logical; if TRUE, the moderator is treated as continuous; if FALSE, it is treated as a factor.

other_covariates
Character vector that specifies the column name of background covariates. Note: Categorical variables must be passed as factors via data.

random_slopes
Logical; if TRUE, the slopes for the main predictor are allowed to vary by cell.

re_parties
Logical; if TRUE, higher-level cell-specific intercepts at the level of voteswitchR::mappings$party_harmonized are added.

re_elections
Logical; if TRUE, higher-level cell-specific intercepts at the level of voteswitchR::mappings$elec_id are added. Note: Should only be used when studying party-electorates.

re_countries
Logical; if TRUE, higher-level cell-specific intercepts at the level of voteswitchR::mappings$iso2c are added.

iter
A positive integer specifying the number of iterations for each chain (including warmup). The default is 2000.

warmup
A positive integer specifying the number of warmup (aka burnin) iterations per chain. If step-size adaptation is on (which it is by default), this also controls the number of iterations for which adaptation is run (and hence these warmup samples should not be used for inference). The number of warmup iterations should be smaller than iter and the default is iter/2.

thin
A positive integer specifying the period for saving samples. The default is 1, which is usually the recommended value.

chains
A positive integer specifying the number of Markov chains. The default is 4.

cores
Number of cores to use when executing the chains in parallel, which defaults to 1 but we recommend setting the mc.cores option to be as many processors as the hardware and RAM allow (up to the number of chains).

seed
The seed for random number generation. The default is generated from 1 to the maximum integer supported by R on the machine. Even if multiple chains are used, only one seed is needed, with other chains having seeds derived from that of the first chain to avoid dependent samples. When a seed is specified by a number, as.integer will be applied to it. If as.integer produces NA, the seed is generated randomly. The seed can also be specified as a character string of digits, such as "12345", which is converted to integer.

parallelize
Logical; if TRUE, estimation is not only parallelized within imputations but concurrently across imputations. Note: This initializes many processes and is only recommended if sufficient cores are available.

savename
A character that specifies the file name of a log file.

Value

Returns an object of class mavcl_est: A list that includes estimates and auxiliary information.

[Package voteswitchR version 0.1.0 Index]